requirements.txt
django-codemirror-widget
